requirements.txt
aiofiles==24.1.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.5
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.1
dataclasses-json==0.6.7
emoji==2.14.1
faiss-cpu==1.12.0
fastapi==0.116.2
filelock==3.19.1
filetype==1.2.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.1.10
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain-core==0.3.76
langdetect==1.0.9
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.8
langsmith==0.4.29
lxml==6.0.1
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
mypy_extensions==1.1.0
narwhals==2.5.0
networkx==3.5
nltk==3.9.1
numpy==2.3.3
olefile==0.47
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
protobuf==6.32.1
psutil==7.1.0
pyarrow==21.0.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
PyMuPDF==1.26.4
pypdf==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.2
rank-bm25==0.2.2
RapidFuzz==3.14.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
streamlit==1.49.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.0
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
unstructured==0.18.15
unstructured-client==0.42.3
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.3
xxhash==3.5.0
zstandard==0.25.0
anthropic>=0.34.0